% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingStars.R
\name{html_dependency_applause}
\alias{html_dependency_applause}
\title{Create an HTML dependency for star rating}
\usage{
html_dependency_applause()
}
\value{
An HTML dependency object.
}
\description{
This function creates an HTML dependency for star rating using specified scripts and stylesheets.
}
