% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingStars.R
\name{html_stmt}
\alias{html_stmt}
\title{Create HTML statement for rating input}
\usage{
html_stmt(h_id, h_label)
}
\arguments{
\item{h_id}{The ID for the input element.}

\item{h_label}{The label for the input element.}
}
\value{
HTML code for rating input
}
\description{
This function creates HTML code for a rating input element.
}
