% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingStars.R
\name{ratingInput}
\alias{ratingInput}
\title{Create a rating input widget}
\usage{
ratingInput(id, label_name)
}
\arguments{
\item{id}{The ID for the input element.}

\item{label_name}{The label name for the input element.}
}
\value{
An HTML widget for rating input.
}
\description{
This function creates a rating input widget using HTML and CSS, enhanced with JavaScript.
}
\examples{
ratingInput("ratingId", "Rating")

}
